requirements.txt
langgraph
langchain
langchain_core
langchain_ollama
langchain_huggingface
langchain_qdrant
sentence-transformers
pillow
qdrant_client
python-dotenv
fastapi
PyMuPDF
uvicorn